requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
sqlalchemy==1.4.22
mysql-connector-python==8.0.27
requests==2.26.0
pytest==7.2.2
pydantic_settings